Move most code from op cuda to source file
